CloudFunctions/upload_to_bigquery/requirements.txt
google-cloud-bigquery
google-cloud-storage
